+++ truck-shapeops/src/transversal/mod.rs
@@ -1,7 +1,8 @@
 mod divide_face;
 mod faces_classification;
 mod integrate;
-mod intersection_curve;
+pub(crate) mod intersection_curve;
 mod loops_store;
 mod polyline_construction;
+pub(crate) use integrate::smooth_leader;
 pub use integrate::{and, or, ShapeOpsCurve, ShapeOpsSurface};